vision/requirements.txt
mediapipe==0.10.14
opencv-python==4.10.0.84
numpy==1.26.4
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pyyaml==6.0.1
pydantic==2.5.0